Split up files in preparation of api stubbing
