+++ lib/ringcaptcha/api_stub.rb
@@ -1,0 +1,9 @@
+require 'json'
+
+module Ringcaptcha
+  module APIStub
+    def self.api(app_key, path, params = {})
+      return Response.new(json.symbolize_keys!)
+    end
+  end
+end
