+++ lib/ringcaptcha/response.rb
@@ -1,0 +1,13 @@
+require 'ostruct'
+
+module Ringcaptcha
+  class Response < OpenStruct
+    def error?
+      self.status == "ERROR"
+    end
+
+    def success?
+      self.status == "SUCCESS"
+    end
+  end
+end
